Omaha #3477 changed plot info locations to floats

Former-commit-id: 73b2585d802d02fd9e20e2acbbf79e39ca5476b3 [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426] [formerly f8a1e79288acbe72e876c449f3ef764397c0743d] [formerly 73b2585d802d02fd9e20e2acbbf79e39ca5476b3 [formerly 3b97fd15a7a3f5d293859ed086844af660ee9426] [formerly f8a1e79288acbe72e876c449f3ef764397c0743d] [formerly 8738e9cecec47e090abd6822026187e958b357f7 [formerly f8a1e79288acbe72e876c449f3ef764397c0743d [formerly 00504d6aa7fefd7b6d28701aae777c0fb7d5ace5]]]]
Former-commit-id: 8738e9cecec47e090abd6822026187e958b357f7
Former-commit-id: 393659ece82f58e3e33b9925c3638c440298b48e [formerly 90a17fa5a6b0c35df35802357453819680aac2f9] [formerly cc959db02bdac9bc65cd7a6ddad09207c1ce1224 [formerly a180c212c921a6a0bbeb490045c05185d2207ef3]]
Former-commit-id: c98315075490f91f8cf521ed63c62b0940c10bfc [formerly fd5c0ea48158e90826270be92bad5eae3b9ca93e]
Former-commit-id: 8b8ad6e45e070c04f3394c37f2e8096340f841ac

+++ ncep/gov.noaa.nws.ncep.viz.rsc.plotdata/src/gov/noaa/nws/ncep/viz/rsc/plotdata/rsc/NcPlotResource2.java
@@ -130,6 +130,7 @@ import com.vividsolutions.jts.geom.Coordinate;
  *  05/20/2013     988     Archana.S   Refactored this class for performance improvement	
  *  11/07/2013             sgurung     Added fix for "no data for every other frame" issue (earlier fix was added to 13.5.2 on 10/24/2013)
  *  03/18/2013    1064     B. Hebbard  Added handling of matrixType request constraint, for PAFM
+ *  Aug 08, 2014  3477     bclement    changed plot info locations to floats
  * </pre>
  * 
  * @author brockwoo
@@ -264,8 +265,8 @@ public class NcPlotResource2 extends
                         synchronized (stationList) {
                             for (Station station : stationList) {
                                 String stnKey = getStationMapKey(
-                                        station.info.latitude.doubleValue(),
-                                        station.info.longitude.doubleValue());
+                                        station.info.latitude,
+                                        station.info.longitude);
 
                                 if (trialKey == null)
                                     trialKey = new String(stnKey);
@@ -519,10 +520,8 @@ public class NcPlotResource2 extends
              * stored in the database with missing/invalid lat/lon values
              */
 
-            if (plotInfo.latitude.doubleValue() < -90
-                    || plotInfo.latitude.doubleValue() > 90
-                    || plotInfo.longitude.doubleValue() < -180
-                    || plotInfo.longitude.doubleValue() > 180) {
+            if (plotInfo.latitude < -90 || plotInfo.latitude > 90
+                    || plotInfo.longitude < -180 || plotInfo.longitude > 180) {
                 return false;
             }
 
@@ -547,8 +546,8 @@ public class NcPlotResource2 extends
                         out.println("2 stations " + stn.info.stationId
                                 + " and " + plotInfo.stationId
                                 + " have the same location?" + "\nLat = "
-                                + stn.info.latitude.doubleValue() + ",Lon = "
-                                + stn.info.longitude.doubleValue()
+                                + stn.info.latitude + ",Lon = "
+                                + stn.info.longitude
                                 + " for the time: " + stn.info.dataTime);
                     }
                     // if these are the same time, should we check which one
@@ -1612,8 +1611,8 @@ public class NcPlotResource2 extends
         fd.setOfStationsLastRendered.addAll(collectionOfStationsToBeRendered);
 
         for (Station stn : collectionOfStationsToBeRendered) {
-            String stnKey = getStationMapKey(stn.info.latitude.doubleValue(),
-                    stn.info.longitude.doubleValue());
+            String stnKey = getStationMapKey(stn.info.latitude,
+                    stn.info.longitude);
             synchronized (fd.stationMap) {
                 fd.stationMap.put(stnKey, stn);
             }
@@ -1689,7 +1688,7 @@ public class NcPlotResource2 extends
 
     // generate a string used as the key for the StationMap
 
-    private String getStationMapKey(Double lat, Double lon) {
+    private String getStationMapKey(Float lat, Float lon) {
         return new String("" + Math.round(lat * 1000.0) + ","
                 + Math.round(lon * 1000.0));
     }
